Fix "inline" where "static" is intended.
Using "inline" in a .cpp file is rarely correct; it tells the compiler
the function has the same definition in every compilation unit, which is
unlikely to be true unless the same code is being #included via a header
file.

Instead, to tell the compiler this copy is only used in this compilation
unit and not needed elsewhere, the right approach is to use internal
linkage, which is what the "static" keyword does.